Create seperate file for the view model
